requirements-dev.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
boto3==1.13.26
botocore==1.16.26
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-haystack==3.0b2
django-js-asset==1.2.2
django-machina==1.1.1
django-modelcluster==5.0.1
django-mptt==0.11.0
django-storages==1.9.1
django-taggit==1.2.0
django-tailwind==0.7.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
jmespath==0.10.0
l18n==2018.5
markdown2==2.3.9
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.9
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
xlrd==1.2.0
XlsxWriter==1.2.8
